requirements.txt
discord.py 
beautifulsoup4 
requests 
